# Add Sympy generator for RK stepper Jacobian