@stephenswat 2025-08-26
Depends on #1057. Pinding @andiwand.

Before (#1057):

```
------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                    Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------------------
BM_PROPAGATION_TOY_DETECTOR_W_COV_TRANSPORT_262144_TRACKS/real_time   31180615 ns     31078208 ns           23 TracksPropagated=8.40727M/s
BM_PROPAGATION_TOY_DETECTOR_262144_TRACKS/real_time                   14613010 ns     14563836 ns           48 TracksPropagated=17.9391M/s
BM_PROPAGATION_WIRE_CHAMBER_W_COV_TRANSPORT_262144_TRACKS/real_time   31582959 ns     31482697 ns           22 TracksPropagated=8.30017M/s
BM_PROPAGATION_WIRE_CHAMBER_262144_TRACKS/real_time                   13011506 ns     12966675 ns           54 TracksPropagated=20.1471M/s
```

After:

```
------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                    Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------------------
BM_PROPAGATION_TOY_DETECTOR_W_COV_TRANSPORT_262144_TRACKS/real_time   30690268 ns     30437394 ns           23 TracksPropagated=8.5416M/s
BM_PROPAGATION_TOY_DETECTOR_262144_TRACKS/real_time                   14633235 ns     14523274 ns           48 TracksPropagated=17.9143M/s
BM_PROPAGATION_WIRE_CHAMBER_W_COV_TRANSPORT_262144_TRACKS/real_time   31073522 ns     30949117 ns           23 TracksPropagated=8.43625M/s
BM_PROPAGATION_WIRE_CHAMBER_262144_TRACKS/real_time                   12884766 ns     12835030 ns           54 TracksPropagated=20.3453M/s
```